requirements.txt
torch==2.6.0
torchvision==0.21.0
streamlit
nibabel
numpy
pillow
scipy